zeoverify-3.0/ai-engine/requirements.txt
pandas
numpy
scikit-learn
torch
transformers
opencv-python
pytesseract
matplotlib
seaborn
python-docx
pdf2image
Pillow
easyocr
